requirements.txt
Django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.3.1